requirements.txt
Pillow>=10.0.0
nbtlib>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
tqdm>=4.66.0
